Now showing an example of how to mix c++ and blueprints
Specifically, this starts with a c++ class, and then "converts" an *instance* of the c++ class (with all of the initial instance's meshes). This is not really a conversion, but rather creating the blueprint as a child-class to the original C++ class. Notably, this means that changes to the underlying C++ class will still affect the blueprint, even after "conversion".

The C++ class implements an animation that that requires relatively short c++ code but require dozens of blueprint nodes.

The blueprint class calls that behaviour, and scales the speed of the animation with only a handful of nodes.